Customer and Admin pages rough draft
Built out the shell of the customer and admin pages respectively. I left
in Alexey's Java code for now, but I am not too sure how necessary it
will be. Once DB is connected, we can see what makes the most sense for
pulling in data into the web pages.